+++ spec/entry_spec.rb
@@ -9,10 +9,37 @@ describe Chchchanges::Entry do
 
   describe '#call' do
     it 'creates a .changelog_entries directory if it does not exist' do
-      setup_fake_input('quit')
+      allow(File).to receive(:write)
+      setup_fake_input('1.0.0', '1', '001', 'test description', 'test1, test2')
       allow(Dir).to receive(:exists?).with('.changelog_entries').and_return(false)
       expect(Dir).to receive(:mkdir).with('.changelog_entries')
-      Chchchanges::Entry.call
+      subject.call
     end
+
+    context 'user inputs valid data' do
+      before(:example) { allow(Dir).to receive(:exists?).with('.changelog_entries').and_return(true) }
+      it 'creates a new .json file with user entered data' do
+        setup_fake_input('1.0.0', '1', '001', 'test description', 'test1, test2')
+        info = {
+          type: 'Added',
+          ticket: '001',
+          url: "",
+          description: 'test description',
+          version: '1.0.0',
+          tags: ['test1', 'test2']
+        }.to_json
+        expect(File).to receive(:write).with(match(/\.changelog_entries\/\d{17}.+\.json/), info)
+        subject.call
+      end
+    end
+
+    context 'invalid user input' do
+      before(:example) { allow(Dir).to receive(:exists?).with('.changelog_entries').and_return(true) }
+      skip 'displays an error message to user on invalid version' do
+        setup_fake_input('1...23', 'q')
+        expect{ subject.call }.to output(/Invalid version number/).to_stdout
+      end
+    end
+
   end
 end